Add files via upload
Problem Statement:
We all have been in that moment where we are authoring a huge essay and want to make it more formal with all those extravagant words. Human always have the tendency to be do things faster rather than trying to take time and check to proofread. The same way we also hesitate to open a huge dictionary and search all terms and synonyms. So, in order to make this work easier we have come up with a solution.

Solution:
We have come up with an idea of creating a typing assistant that is available as a browser extension in any browser that people wish to use. This typing assistant keeps running behind your browser whenever you are editing and authoring your paper. It gives suggestion on which sentences are good and which sentences are more formal and informal. It also has the ability to translate any languages to the one that the user prefers. It acts as a dictionary that automatically searches the word on the internet and displays the search results right on the tab that you are typing out on.